refactor add pageState directory
