+++ src/analysis/analysisVisitors/pageStateAnalysis/pageStateAnalysis.test.ts
@@ -16,7 +16,10 @@
  */
 
 import { formStateFactory } from "@/testUtils/factories/pageEditorFactories";
-import { GetPageState, SetPageState } from "@/bricks/effects/pageState";
+import {
+  GetPageState,
+  SetPageState,
+} from "@/bricks/effects/pageState/pageState";
 import PageStateAnalysis from "@/analysis/analysisVisitors/pageStateAnalysis/pageStateAnalysis";
 import { AnnotationType } from "@/types/annotationTypes";
 import { registryIdFactory } from "@/testUtils/factories/stringFactories";
