+++ src/bricks/effects/pageState/pageState.test.ts
@@ -28,7 +28,9 @@ beforeEach(() => {
 
 describe("@pixiebrix/state/get", () => {
   test("default to blueprint state", async () => {
-    const { GetPageState } = await import("@/bricks/effects/pageState");
+    const { GetPageState } = await import(
+      "@/bricks/effects/pageState/pageState"
+    );
 
     const brick = new GetPageState();
     const logger = new ConsoleLogger({
@@ -39,7 +41,9 @@ describe("@pixiebrix/state/get", () => {
   });
 
   test("is page state aware", async () => {
-    const { GetPageState } = await import("@/bricks/effects/pageState");
+    const { GetPageState } = await import(
+      "@/bricks/effects/pageState/pageState"
+    );
     const brick = new GetPageState();
     await expect(brick.isPageStateAware()).resolves.toBe(true);
   });
@@ -47,7 +51,9 @@ describe("@pixiebrix/state/get", () => {
 
 describe("@pixiebrix/state/set", () => {
   test("shallow merge", async () => {
-    const { SetPageState } = await import("@/bricks/effects/pageState");
+    const { SetPageState } = await import(
+      "@/bricks/effects/pageState/pageState"
+    );
 
     const brick = new SetPageState();
     const logger = new ConsoleLogger({
@@ -71,7 +77,9 @@ describe("@pixiebrix/state/set", () => {
   });
 
   test("deep merge does not append array elements", async () => {
-    const { SetPageState } = await import("@/bricks/effects/pageState");
+    const { SetPageState } = await import(
+      "@/bricks/effects/pageState/pageState"
+    );
 
     const brick = new SetPageState();
     const logger = new ConsoleLogger({
@@ -118,7 +126,9 @@ describe("@pixiebrix/state/set", () => {
   });
 
   it("returns mod variables", async () => {
-    const { SetPageState } = await import("@/bricks/effects/pageState");
+    const { SetPageState } = await import(
+      "@/bricks/effects/pageState/pageState"
+    );
     const brick = new SetPageState();
 
     await expect(
@@ -141,7 +151,9 @@ describe("@pixiebrix/state/set", () => {
   });
 
   it("ignores private variables", async () => {
-    const { SetPageState } = await import("@/bricks/effects/pageState");
+    const { SetPageState } = await import(
+      "@/bricks/effects/pageState/pageState"
+    );
     const brick = new SetPageState();
 
     await expect(
@@ -161,7 +173,7 @@ describe("@pixiebrix/state/set", () => {
 describe("set and get", () => {
   test("default to blueprint state", async () => {
     const { GetPageState, SetPageState } = await import(
-      "@/bricks/effects/pageState"
+      "@/bricks/effects/pageState/pageState"
     );
 
     const setState = new SetPageState();
@@ -197,7 +209,7 @@ describe("set and get", () => {
 
   test("default to shared if not part of blueprint", async () => {
     const { GetPageState, SetPageState } = await import(
-      "@/bricks/effects/pageState"
+      "@/bricks/effects/pageState/pageState"
     );
 
     const setState = new SetPageState();
@@ -231,7 +243,9 @@ describe("set and get", () => {
   });
 
   test("is page state aware", async () => {
-    const { SetPageState } = await import("@/bricks/effects/pageState");
+    const { SetPageState } = await import(
+      "@/bricks/effects/pageState/pageState"
+    );
     const brick = new SetPageState();
     await expect(brick.isPageStateAware()).resolves.toBe(true);
   });
